setting up preview on transactionRow
