+++ ccSchwabManager/Views/HoldingsView/PositionDetailView/PositionDetailContent/TransactionsTab/TransactionsTab.swift
@@ -17,4 +17,40 @@ struct TransactionsTab: View {
             Label("Transactions", systemImage: "list.bullet")
         }
     }
+}
+
+#Preview("TransactionsTab - With Data", traits: .landscapeLeft) {
+    GeometryReader { geometry in
+        TransactionsTab(
+            isLoading: false,
+            symbol: "AAPL",
+            geometry: geometry
+        )
+    }
+    .frame(width: 800, height: 600)
+    .padding()
+}
+
+#Preview("TransactionsTab - Loading State", traits: .landscapeLeft) {
+    GeometryReader { geometry in
+        TransactionsTab(
+            isLoading: true,
+            symbol: "AAPL",
+            geometry: geometry
+        )
+    }
+    .frame(width: 800, height: 600)
+    .padding()
+}
+
+#Preview("TransactionsTab - Different Symbol", traits: .landscapeLeft) {
+    GeometryReader { geometry in
+        TransactionsTab(
+            isLoading: false,
+            symbol: "MSFT",
+            geometry: geometry
+        )
+    }
+    .frame(width: 800, height: 600)
+    .padding()
 }
